+++ app/domains/fluggastrechte/services/pdf/sections/firstPage/claimData/addPlaintiffDetails.ts
@@ -3,8 +3,8 @@ import type { FluggastrechtContext } from "~/domains/fluggastrechte/formular/con
 import {
   FONTS_BUNDESSANS_BOLD,
   FONTS_BUNDESSANS_REGULAR,
-} from "~/services/pdf/fluggastrechte/createPdfKitDocument";
-import { getFullPlaintiffName } from "~/services/pdf/fluggastrechte/sections/getFullPlaintiffName";
+} from "../../../createPdfKitDocument";
+import { getFullPlaintiffName } from "../../getFullPlaintiffName";
 
 export const PLAINTIFF_TEXT = "– Klagende Partei –";
 export const SEPARATOR = " | ";
