+++ app/domains/fluggastrechte/services/pdf/sections/reason/factsOfCases/__test__/addDetailedReason.test.ts
@@ -4,7 +4,7 @@ import {
   mockPdfKitDocumentStructure,
 } from "tests/factories/mockPdfKit";
 import { getAirportNameByIataCode } from "~/services/airports/getAirportNameByIataCode";
-import { PDF_MARGIN_HORIZONTAL } from "~/services/pdf/fluggastrechte/createPdfKitDocument";
+import { PDF_MARGIN_HORIZONTAL } from "../../../../createPdfKitDocument";
 import {
   addDetailedReason,
   ATTACHMENT_CONFIRM_BOOKING_TEXT,
