+++ app/domains/fluggastrechte/services/pdf/sections/reason/factsOfCases/table/drawCell.ts
@@ -2,7 +2,7 @@ import type PDFDocument from "pdfkit";
 import {
   FONTS_BUNDESSANS_BOLD,
   FONTS_BUNDESSANS_REGULAR,
-} from "~/services/pdf/fluggastrechte/createPdfKitDocument";
+} from "../../../../createPdfKitDocument";
 
 type CellOptions = {
   xPosition: number;
