+++ app/domains/fluggastrechte/services/pdf/sections/reason/factsOfCases/table/drawTableColumnsValues.ts
